refactor: rename helpers to utils
